Fix DateTimePicker tests
