+++ lib/src/__tests__/e2e/DateTimePickerKeyboard.test.tsx
@@ -1,0 +1,48 @@
+import { ReactWrapper } from 'enzyme';
+import * as React from 'react';
+import KeyboardDateTimePicker, {
+  KeyboardDateTimePickerProps,
+} from '../../DateTimePicker/KeyboardDateTimePicker';
+import { mount, utilsToUse } from '../test-utils';
+
+const format = process.env.UTILS === 'moment' ? 'MM/DD/YYYY HH:mm' : 'MM/dd/yyyy hh:mm';
+
+describe('e2e - KeyboardDateTimePicker (inline)', () => {
+  let component: ReactWrapper<KeyboardDateTimePickerProps>;
+  const onChangeMock = jest.fn();
+  const onCloseMock = jest.fn();
+  const onOpenMock = jest.fn();
+
+  beforeEach(() => {
+    jest.clearAllMocks();
+    component = mount(
+      <KeyboardDateTimePicker
+        // @ts-ignore
+        variant="inline"
+        onChange={onChangeMock}
+        onClose={onCloseMock}
+        onOpen={onOpenMock}
+        format={format}
+        KeyboardButtonProps={{ id: 'keyboard-button' }}
+        value={utilsToUse.date('2018-01-01T00:00:00.000Z')}
+      />
+    );
+  });
+
+  it('Should renders', () => {
+    expect(component).toBeTruthy();
+  });
+
+  it('Should open modal with picker on click', () => {
+    component.find('button#keyboard-button').simulate('click');
+    expect(component.find('WithStyles(Popover)').props().open).toBeTruthy();
+    // TODO Support onOpen
+    // expect(onOpenMock).toHaveBeenCalled();
+  });
+
+  it('Should apply keyboard input', () => {
+    component.find('input').simulate('change', { target: { value: '02/01/2018 10:00' } });
+
+    expect(onChangeMock).toHaveBeenCalled();
+  });
+});
