+++ lib/src/_shared/PureDateInput.tsx
@@ -31,3 +31,5 @@ export const PureDateInput: React.FC<PureDateInputProps> = ({
     />
   );
 };
+
+PureDateInput.displayName = 'PureDateInput';
